Day 1 submission
Base it off on a priority queue.
Started it based on a dictionary but since the requirements just needed a quick calculation, I performed the summation before adding it to the queue. 🤷🏽‍♂️
That way, when it came to getting the bag with most calories, it was a matter of dequeuing a pre-calculated value.